demo/requirements.txt
bleak==0.20.2
bleak-winrt==1.2.0
contourpy==1.1.1
cycler==0.12.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.1
Pillow==10.0.1
pynput==1.7.6
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
